edit validatioin logic | cleanup
