+++ src/applications/edu-benefits/0839/hooks/useValidateFacilityCode.js
@@ -6,16 +6,9 @@ import { setData } from 'platform/forms-system/src/js/actions';
 export const useValidateFacilityCode = formData => {
   const [loader, setLoader] = useState(false);
   const [institutionData, setInstitutionData] = useState(null);
-  // const [lastFetchedCode, setLastFetchedCode] = useState(null);
   const dispatch = useDispatch();
 
   const facilityCode = formData?.institutionDetails?.facilityCode;
-
-  // const rawFacilityCode = formData?.institutionDetails?.facilityCode;
-  // const facilityCode =
-  //   typeof rawFacilityCode === 'string'
-  //     ? rawFacilityCode.trim()
-  //     : rawFacilityCode?.toString()?.trim() || '';
 
   useEffect(
     () => {
@@ -109,14 +102,8 @@ export const useValidateFacilityCode = formData => {
       if (facilityCode?.length === 8) {
         fetchInstitutionInfo();
       }
-      // if (facilityCode?.length === 8 && facilityCode !== lastFetchedCode) {
-      //   console.log({ facilityCode, lastFetchedCode });
-      //   setLastFetchedCode(facilityCode);
-      //   fetchInstitutionInfo();
-      // }
     },
-    [facilityCode, formData, dispatch],
-    // [facilityCode, formData, dispatch, lastFetchedCode],
+    [facilityCode],
   );
   const attrs = institutionData?.attributes || {};
   const institutionAddress = {
